requirements.txt
annotated-types==0.6.0
anyio==4.3.0
azure-core==1.30.1
azure-identity==1.15.0
azure-storage-blob==12.19.1
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
fastapi==0.110.0
idna==3.6
isodate==0.6.1
msal==1.27.0
msal-extensions==1.1.0
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==23.2
portalocker==2.8.2
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
PyMySQL==1.1.0
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1